Split encoding.py to schemas and fields
