+++ rotkehlchen/api/v1/resources.py
@@ -13,7 +13,8 @@ from werkzeug.datastructures import FileStorage
 from rotkehlchen.accounting.ledger_actions import LedgerAction, LedgerActionType
 from rotkehlchen.accounting.structures import ActionType, HistoryBaseEntry
 from rotkehlchen.api.rest import RestAPI
-from rotkehlchen.api.v1.encoding import (
+from rotkehlchen.api.v1.parser import ignore_kwarg_parser, resource_parser
+from rotkehlchen.api.v1.schemas import (
     AccountingReportDataSchema,
     AccountingReportsSchema,
     AllBalancesQuerySchema,
@@ -110,7 +111,6 @@ from rotkehlchen.api.v1.encoding import (
     XpubAddSchema,
     XpubPatchSchema,
 )
-from rotkehlchen.api.v1.parser import ignore_kwarg_parser, resource_parser
 from rotkehlchen.assets.asset import Asset, EthereumToken
 from rotkehlchen.assets.typing import AssetType
 from rotkehlchen.balances.manual import ManuallyTrackedBalance
