+++ rotkehlchen/chain/ethereum/decoding/uniswap/v1/decoder.py
@@ -39,7 +39,7 @@ class Uniswapv1Decoder(DecoderInterface):
             token: Optional[EthereumToken],  # pylint: disable=unused-argument
             tx_log: EthereumTxReceiptLog,
             transaction: EthereumTransaction,  # pylint: disable=unused-argument
-            decoded_events: List[HistoryBaseEntry],  # pylint: disable=unused-argument
+            decoded_events: List[HistoryBaseEntry],
             action_items: List[ActionItem],  # pylint: disable=unused-argument
     ) -> Optional[HistoryBaseEntry]:
         if tx_log.topics[0] == TOKEN_PURCHASE:
